# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StockTank API",
    "version": "0.1.0",
    "description": "Identity, RBAC, public content, advertising, marketing and system endpoints for StockTank. Sessions are cookie-based; state-changing requests must send `X-Requested-With`. Financial content served by this platform is informational only and never investment advice."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin (dev proxy or deployed API host)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "st_session",
        "description": "httpOnly session cookie set by /api/v1/auth/register and /api/v1/auth/login. The server stores only a SHA-256 hash of the token."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "VALIDATION_FAILED",
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "INTERNAL"
                ]
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ],
        "description": "Envelope used by every non-2xx response.",
        "example": {
          "error": {
            "code": "VALIDATION_FAILED",
            "message": "Invalid request body",
            "requestId": "7f1c1b7e-2d1a-4a4d-9d2b-3b7b2a1f0c9e",
            "details": {
              "source": "body",
              "issues": [
                {
                  "path": "email",
                  "message": "Invalid email address",
                  "code": "invalid_format"
                }
              ]
            }
          }
        }
      },
...
      "MediaKitResponse": {
        "type": "object",
        "properties": {
          "placements": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string",
                  "enum": [
                    "home_presenting_sponsor",
                    "home_native_feed",
                    "home_leaderboard",
                    "watchlist_sidebar",
                    "episode_page",
                    "project_page",
                    "company_page",
                    "newsletter_primary",
                    "audio_preroll",
                    "video_preroll",
                    "show_sponsorship"
                  ]
                },
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "surface": {
                  "type": "string",
                  "enum": [
                    "web",
                    "newsletter",
                    "audio",
                    "video"
                  ]
                },
                "format": {
                  "type": "string",
                  "enum": [
                    "sponsor_banner",
                    "native_card",
                    "leaderboard",
                    "sidebar_card",
                    "audio_read",
                    "video_preroll",
                    "show_sponsorship",
                    "newsletter_slot"
                  ]
                },
                "specs": {
                  "type": "string"
                },
                "pricingModel": {
                  "type": "string",
                  "enum": [
                    "cpm",
                    "flat_week",
                    "flat_episode",
                    "flat_issue"
                  ]
                },
                "rateCents": {
                  "type": [
                    "integer",
                    "null"
                  ]
                },
                "currency": {
                  "type": "string"
                },
                "rateVisibility": {
                  "type": "string",
                  "enum": [
                    "public",
                    "on_request"
                  ]
                }
              },
              "required": [
                "key",
                "name",
                "description",
                "surface",
                "format",
                "specs",
                "pricingModel",
                "rateCents",
                "currency",
                "rateVisibility"
              ]
            }
          },
          "advertisingLive": {
            "type": "boolean"
          }
        },
        "required": [
          "placements",
          "advertisingLive"
        ]
      },
      "ServeAdResponse": {
        "type": "object",
        "properties": {
          "ad": {
            "type": [
              "object",
              "null"
            ],
            "properties": {
              "placement": {
                "type": "string",
                "enum": [
                  "home_presenting_sponsor",
                  "home_native_feed",
                  "home_leaderboard",
                  "watchlist_sidebar",
                  "episode_page",
                  "project_page",
                  "company_page",
                  "newsletter_primary",
                  "audio_preroll",
                  "video_preroll",
                  "show_sponsorship"
                ]
              },
              "kind": {
                "type": "string",
                "enum": [
                  "display",
                  "native",
                  "audio_script",
                  "video"
                ]
              },
              "advertiserName": {
                "type": "string"
              },
              "headline": {
                "type": "string"
              },
              "body": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "imageUrl": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "altText": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "ctaLabel": {
                "type": "string"
              },
              "clickUrl": {
                "type": "string"
              },
              "disclosureLabel": {
                "type": "string"
              },
              "impressionToken": {
                "type": "string"
              },
              "isHouse": {
                "type": "boolean"
              }
            },
            "required": [
              "placement",
              "kind",
              "advertiserName",
              "headline",
              "body",
              "imageUrl",
              "altText",
              "ctaLabel",
              "clickUrl",
              "disclosureLabel",
              "impressionToken",
              "isHouse"
            ]
          }
        },
        "required": [
          "ad"
        ]
      },
      "AdvertisingInquiryRequest": {
        "type": "object",
        "properties": {
          "company": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "contactName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email"
          },
          "website": {
            "type": "string",
            "maxLength": 2048,
            "format": "uri"
          },
          "budgetRange": {
            "type": "string",
            "enum": [
              "under_5k",
              "from_5k_to_25k",
              "from_25k_to_100k",
              "over_100k",
              "undisclosed"
            ]
          },
          "placementKeys": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "home_presenting_sponsor",
                "home_native_feed",
                "home_leaderboard",
                "watchlist_sidebar",
                "episode_page",
                "project_page",
                "company_page",
                "newsletter_primary",
                "audio_preroll",
                "video_preroll",
                "show_sponsorship"
              ]
            },
            "maxItems": 11,
            "default": []
          },
          "message": {
            "type": "string",
            "minLength": 10,
            "maxLength": 2000
          },
          "consent": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "companyFax": {
            "type": "string",
            "maxLength": 500
          },
          "utm": {
            "type": "object",
            "properties": {
              "source": {
                "type": "string",
                "maxLength": 100
              },
              "medium": {
                "type": "string",
                "maxLength": 100
              },
              "campaign": {
                "type": "string",
                "maxLength": 100
              }
            }
          },
          "referrer": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "company",
          "contactName",
          "email",
          "budgetRange",
          "message",
          "consent"
        ],
        "example": {
          "company": "Acme Custody",
          "contactName": "Pat Doe",
          "email": "pat@acme.example",
          "budgetRange": "from_5k_to_25k",
          "placementKeys": [
            "newsletter_primary"
          ],
          "message": "We would like to sponsor the RWA Report next quarter.",
          "consent": true
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/advertising/media-kit": {
      "get": {
        "tags": [
          "Advertising"
        ],
        "summary": "Media kit",
        "description": "Active placements with specs. Rates appear only when sales has published them.",
        "responses": {
          "200": {
            "description": "Placements",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaKitResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ads/serve": {
      "get": {
        "tags": [
          "Advertising"
        ],
        "summary": "Serve an ad for a placement",
        "description": "Returns an approved, in-flight ad or null. Sets a first-party `st_vid` visitor cookie for frequency capping. `clickUrl` is always the tracking URL; `impressionToken` is signed and single-use.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "home_presenting_sponsor",
                "home_native_feed",
                "home_leaderboard",
                "watchlist_sidebar",
                "episode_page",
                "project_page",
                "company_page",
                "newsletter_primary",
                "audio_preroll",
                "video_preroll",
                "show_sponsorship"
              ]
            },
            "required": true,
            "name": "placement",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 300
            },
            "required": false,
            "name": "path",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "An ad, or { ad: null }",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServeAdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ads/impressions": {
      "post": {
        "tags": [
          "Advertising"
        ],
        "summary": "Record a viewable impression",
        "description": "Idempotent per token.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "CSRF guard required on every state-changing request under /api. Any non-empty value (the official clients send \"stocktank\").",
              "example": "stocktank"
            },
            "required": true,
            "description": "CSRF guard required on every state-changing request under /api. Any non-empty value (the official clients send \"stocktank\").",
            "name": "x-requested-with",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 2048
                  }
                },
                "required": [
                  "token"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Recorded (or already recorded)"
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ads/click/{token}": {
      "get": {
        "tags": [
          "Advertising"
        ],
        "summary": "Track a click and redirect",
        "description": "Records the click and redirects (302) to the destination stored on the creative. Never an open redirect.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "token",
            "in": "path"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to the advertiser"
          },
          "404": {
            "description": "Ad link not found (NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/advertising/inquiries": {
      "post": {
        "tags": [
          "Marketing"
        ],
        "summary": "Submit an advertising inquiry",
        "description": "Stores a sales lead with UTM attribution and notifies sales when configured. 5 per hour per IP. Honeypot submissions are accepted and discarded.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "CSRF guard required on every state-changing request under /api. Any non-empty value (the official clients send \"stocktank\").",
              "example": "stocktank"
            },
            "required": true,
            "description": "CSRF guard required on every state-changing request under /api. Any non-empty value (the official clients send \"stocktank\").",
            "name": "x-requested-with",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvertisingInquiryRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Received",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "received"
                      ]
                    }
                  },
                  "required": [
                    "status"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}